Remove caching query service
This is not needed anymore as we now finalize the property on reads.
